+++ sdmx-dl-provider-connectors/src/main/java/internal/sdmxdl/connectors/drivers/Sdmx21Driver.java
@@ -162,6 +162,16 @@ public final class Sdmx21Driver implements SdmxWebDriver {
                     .build())
             .source(SdmxWebSource
                     .builder()
+                    .name("SIMEL")
+                    .description("El Salvador Labour Market Information System")
+                    .driver(CONNECTORS_SDMX_21)
+                    .endpointOf("https://disseminatesimel.mtps.gob.sv/rest")
+                    .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
+                    .websiteOf("https://datasimel.mtps.gob.sv/")
+                    .monitorOf("upptime:/nbbrd/sdmx-upptime/SIMEL")
+                    .build())
+            .source(SdmxWebSource
+                    .builder()
                     .name("SPC")
                     .description("Pacific Data Hub")
                     .driver(CONNECTORS_SDMX_21)